task 5 update

+++ 0x00-hello_world/6-size.c
@@ -15,11 +15,11 @@ int main(void)
 	char c;
 	float f;
 
-	printf("Size of a char: %lu byte(S)\n",(unsigned long)sizeof(c));
-	printf("Size of a int: %lu byte(s)\n",(unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n",(unsigned long)sizeof(li));
-	printf("Size of a long long int: %lu byte(s)\n",(unsigned long)sizeof(lli));
-	printf("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(f));
+	printf("Size of a char: %lu byte(S)\n", sizeof(c));
+	printf("Size of a int: %lu byte(s)\n", sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(li));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(lli));
+	printf("Size of a float: %lu byte(s)\n", sizeof(f));
 
 	return (0);
 }
